// payload-website-builder-poc/src/payload-types.ts
export interface Page {
  id: number;
  title: string;
  heroBackgroundColor?: ('background' | 'muted' | 'primary') | null;
  heroDesignVersion: 'hero231' | 'hero12' | 'hero18';
  heroBadge?: string | null;
  heroTagline: string;
  heroDescription: string;
  heroPrimaryLabel: string;
  heroPrimaryHref: string;
  heroSecondaryLabel?: string | null;
  heroSecondaryHref?: string | null;
  layout: (FeatureBlock | TestimonialBlock | CallToActionBlock | FAQBlock)[];
  metaTitle?: string | null;
  metaDescription?: string | null;
  slug: string;
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FeatureBlock".
 */
export interface FeatureBlock {
  eyebrow?: string | null;
  headline: string;
  description: string;
  items?:
    | {
        title: string;
        description: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'feature';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TestimonialBlock".
 */
export interface TestimonialBlock {
  quote: string;
  name: string;
  role: string;
  stats?:
    | {
        value: string;
        label: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'testimonial';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  eyebrow?: string | null;
  headline: string;
  description: string;
  primaryLabel: string;
  primaryHref: string;
  secondaryLabel?: string | null;
  secondaryHref?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'callToAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FAQBlock".
 */
export interface FAQBlock {
  headline: string;
  items?:
    | {
        question: string;
        answer: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'faq';
}